removes files and classes / protocols
