+++ Sources/Root/FilterViewController.swift
@@ -1,0 +1,9 @@
+//
+//  Copyright © FINN.no AS, Inc. All rights reserved.
+//
+
+import UIKit
+
+public class FilterViewController: UIViewController {
+    public weak var filterSelectionDelegate: FilterSelectionDelegate?
+}
